simplify: dedupe iframe URL helper + share the tab-record parser
Two cleanups from the /simplify review:

- IframePanel: commitUrl and observeFrameUrl differed by a single
  updateParameters line (and repeated the title formatting a third time
  in goToHistoryIndex). Extract applyFrameUrl(url, persist).

- The host's `tab list --json` parser re-implemented the connection's
  parseTabRecord — same tabId/id fallback, same url/active coercion, even
  the same "some CLI builds use id" rationale. Lift the record parse into
  a shared lib/src/lib/agent-browser-tab.ts that both the connection
  (live stream) and the host (CLI) import, so the tab shape has one home.
